feat: [PIPE-22045]: file explorer in gitness
